+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.49.0] - 2023-01-25
+
+### Added
+
+- `pageCategory` var at event tracking
+- `pathCategoryExtractor` option to override default category selection
+
 ## [3.48.1] - 2023-01-25
 
 ### Fixed